+++ package.json
@@ -14,6 +14,7 @@
     "react-leaflet": "^2.2.1",
     "react-map-gl": "^4.1.1",
     "react-redux": "^6.0.1",
+    "react-router-dom": "^5.0.0",
     "react-scripts": "2.1.8",
     "redux": "^4.0.1",
     "redux-thunk": "^2.3.0"
