+++ src/containers/TrackoContainer.js
@@ -1,64 +1,38 @@
 import React, { Component } from 'react';
-import { connect } from 'react-redux';
-import { bindActionCreators } from 'redux';
-import * as actions from '../actions';
+import { BrowserRouter, Route, Link } from 'react-router-dom';
+import TrackoEventsContainer from './TrackoEventsContainer';
 
-import Map from '../components/Map';
-import LeftDrawer from '../components/LeftDrawer';
-import BottomSlider from '../components/BottomSlider';
-import LiveOrRecFab from '../components/LiveOrRecFab';
-import { mapboxConfig } from '../mapbox/config';
 
 class TrackoContainer extends Component {
-	componentDidMount() {
-		const { actions } = this.props;
-		actions.loadClassList();
-		actions.loadUserlist();
-		actions.loadUserLocations();
-	}
-
 	render() {
-		const token = mapboxConfig.token;
-		const { MapReducer, FirebaseDbReducer, LeftDrawerReducer, SettingsReducer, actions } = this.props;
-		// const { MapReducer, FirebaseDbReducer, LeftDrawerReducer, BottomSliderReducer, LiveOrRecFabReducer, actions } = this.props;
-		// console.log(SettingsReducer);
-		if(FirebaseDbReducer.userlist === undefined || FirebaseDbReducer.userlocations === undefined || FirebaseDbReducer.classlist === undefined) {
-			return <div>loading</div>;
-		}
 		return (
-			<div>
-				<LeftDrawer
-				compName="第n回日本学生オリエンテーリング選手権大会"/>
-				<BottomSlider
-					sliderValue={SettingsReducer.sliderValue}
-					onBottomSliderChange={actions.onBottomSliderChange} />
-				<Map
-					viewport={MapReducer.viewport}
-					token={token}
-					onViewportChange={(viewport) => actions.onViewportChange(viewport)}
-					sliderValue={SettingsReducer.sliderValue}
-					users={FirebaseDbReducer.users}
-					userlist={FirebaseDbReducer.userlist}
-					userlocations={FirebaseDbReducer.userlocations}
-					viewallFlags={LeftDrawerReducer.viewallFlags}
-					selectedClass={(LeftDrawerReducer.selectedClass === undefined) ? FirebaseDbReducer.classlist['class1'].name : LeftDrawerReducer.selectedClass}/>
-				<LiveOrRecFab
-					liveFlag={SettingsReducer.liveFlag}
-					onLiveOrRecChange={actions.onLiveOrRecChange}/>
-
-			</div>
+			<BrowserRouter>
+				{/* <ul>
+					<li><Link to='/'>Home</Link></li>
+					<li><Link to='/about'>About</Link></li>
+					<li><Link to='/events'>Friends</Link></li>
+				</ul> */}
+				<div>
+					<Route exact path='/' component={Home} />
+					<Route path='/about' component={About} />
+					<Route path='/events' component={TrackoEventsContainer} />
+				</div>
+			</BrowserRouter>
 		);
 	}
 }
 
-const mapStateToProps = (state) => {
-	return state;
-};
+const Home = () => (
+  <div>
+	  <h2>Home</h2>
+	  <p>Welcome to ようこそ</p>
+  </div>
+)
+const About = () => (
+  <div>
+	  <h2>About</h2>
+	  <p>フレンズに投票するページです</p>
+  </div>
+)
 
-const mapDispatch = (dispatch) => {
-	return {
-		actions: bindActionCreators(actions, dispatch),
-	};
-};
-
-export default connect(mapStateToProps, mapDispatch)(TrackoContainer);
+export default TrackoContainer;
